# backend/app/routers/user_profile.py
"""Endpoints for user profile management."""

import datetime
from typing import cast

from fastapi import APIRouter, Depends, HTTPException, Query
from sqlalchemy import CursorResult, delete, select, update
from sqlalchemy import func as sa_func
from sqlalchemy.ext.asyncio import AsyncSession

from backend.app.auth.dependencies import get_current_user
from backend.app.channel_state import realign_preferred_channel_async
from backend.app.channels import is_bluebubbles_configured, reset_channel_clients
from backend.app.config import (
    resolve_imessage_backend,
    settings,
    update_settings,
)
from backend.app.config_store import (
    get_settings_store,
    strip_unchanged_secrets,
)
from backend.app.database import get_async_db
from backend.app.models import ChannelRoute, HeartbeatLog, LLMUsageLog, User
from backend.app.query_helpers import get_or_404_async
from backend.app.schemas import (
    ChannelConfigResponse,
    ChannelConfigUpdate,
    ChannelRouteListResponse,
    ChannelRouteResponse,
    ChannelRouteUpdate,
    DataSharingConsentRequest,
    DataSharingConsentResponse,
    DeleteHeartbeatLogsResponse,
    HeartbeatLogItemResponse,
    HeartbeatLogListResponse,
    LLMUsageByPurpose,
    LLMUsageSummary,
    ModelConfigResponse,
    ModelConfigUpdate,
    ProviderInfo,
    TelegramBotInfoResponse,
    UserProfileResponse,
    UserProfileUpdate,
)
from backend.app.services.llm_service import get_configured_providers, get_models

router = APIRouter()


def _profile_response(c: User) -> UserProfileResponse:
    return UserProfileResponse(
        id=c.id,
        user_id=c.user_id,
        phone=c.phone,
        timezone=c.timezone,
        soul_text=c.soul_text,
        user_text=c.user_text,
        heartbeat_text=c.heartbeat_text,
        preferred_channel=c.preferred_channel,
        channel_identifier=c.channel_identifier,
        heartbeat_opt_in=c.heartbeat_opt_in,
        heartbeat_frequency=c.heartbeat_frequency,
        heartbeat_max_daily=c.heartbeat_max_daily,
        onboarding_complete=c.onboarding_complete,
        is_active=c.is_active,
        data_sharing_consent=c.data_sharing_consent,
        data_sharing_consent_at=(
            c.data_sharing_consent_at.isoformat() if c.data_sharing_consent_at else None
        ),
        created_at=c.created_at.isoformat(),
        updated_at=c.updated_at.isoformat(),
    )


@router.get("/user/profile", response_model=UserProfileResponse)
async def get_profile(
    current_user: User = Depends(get_current_user),
) -> UserProfileResponse:
    """Return the current user's profile."""
    return _profile_response(current_user)


@router.put("/user/profile", response_model=UserProfileResponse)
async def update_profile(
    body: UserProfileUpdate,
    current_user: User = Depends(get_current_user),
    db: AsyncSession = Depends(get_async_db),
) -> UserProfileResponse:
    """Partial update of the current user's profile."""
    updates = body.model_dump(exclude_unset=True)
    if not updates:
        raise HTTPException(status_code=400, detail="No fields to update")

    # Re-query user in the current session to avoid detached instance issues
    user = await get_or_404_async(db, User, detail="User not found", id=current_user.id)

    for key, value in updates.items():
        setattr(user, key, value)
    await db.commit()
    await db.refresh(user)
    return _profile_response(user)


# ---------------------------------------------------------------------------
# Data sharing consent
# ---------------------------------------------------------------------------
#
# Separate endpoint (not folded into ``PUT /user/profile``) because the
# timestamp must be stamped on every change (opt-in AND opt-out) so
# consent history is reconstructable. Routing through the generic patch
# endpoint would require a special case for this column; cleaner to give
# it its own URL with the invariant in the route signature.


def _data_sharing_consent_now() -> datetime.datetime:
    """Return ``datetime.datetime.now(datetime.UTC)``.

    Pulled out as a helper so tests can monkeypatch the clock when they
    need to verify the timestamp updated to a specific instant. Without
    this seam, asserting "the second PUT bumped the timestamp" can only
    rely on ``>= t1`` ordering, which is non-trivial when both calls
    happen inside one millisecond of test runtime.
    """
    return datetime.datetime.now(datetime.UTC)


@router.get("/user/data-sharing-consent", response_model=DataSharingConsentResponse)
async def get_data_sharing_consent(
    current_user: User = Depends(get_current_user),
) -> DataSharingConsentResponse:
    """Return the current user's data sharing consent state."""
    return DataSharingConsentResponse(
        data_sharing_consent=current_user.data_sharing_consent,
        data_sharing_consent_at=(
            current_user.data_sharing_consent_at.isoformat()
            if current_user.data_sharing_consent_at
            else None
        ),
    )


@router.put("/user/data-sharing-consent", response_model=DataSharingConsentResponse)
async def update_data_sharing_consent(
    body: DataSharingConsentRequest,
    current_user: User = Depends(get_current_user),
    db: AsyncSession = Depends(get_async_db),
) -> DataSharingConsentResponse:
    """Toggle the current user's data sharing consent.

    Stamps ``data_sharing_consent_at`` with the current UTC time on every
    call, regardless of whether the value changed. This makes the column
    a "last toggled at" timestamp rather than a "first opted in at" one,
    which is the cheaper guarantee to keep correct: a one-shot accidental
    double-PUT can't drift the timestamp.
    """
    user = await get_or_404_async(db, User, detail="User not found", id=current_user.id)
    user.data_sharing_consent = body.consent
    user.data_sharing_consent_at = _data_sharing_consent_now()
    await db.commit()
    await db.refresh(user)
    return DataSharingConsentResponse(
        data_sharing_consent=user.data_sharing_consent,
        data_sharing_consent_at=(
            user.data_sharing_consent_at.isoformat() if user.data_sharing_consent_at else None
        ),
    )


# ---------------------------------------------------------------------------
# Channel config
# ---------------------------------------------------------------------------


def _build_channel_config_response() -> ChannelConfigResponse:
    return ChannelConfigResponse(
        telegram_bot_token_set=bool(settings.telegram_bot_token),
        telegram_allowed_chat_id=settings.telegram_allowed_chat_id,
        imessage_backend=resolve_imessage_backend(),
        linq_api_token_set=bool(settings.linq_api_token),
        linq_from_number=settings.linq_from_number,
        linq_allowed_numbers=settings.linq_allowed_numbers,
        linq_preferred_service=settings.linq_preferred_service,
        bluebubbles_configured=is_bluebubbles_configured(),
        bluebubbles_allowed_numbers=settings.bluebubbles_allowed_numbers,
        bluebubbles_imessage_address=settings.bluebubbles_imessage_address,
    )


@router.get("/user/channels/config", response_model=ChannelConfigResponse)
async def get_channel_config(
    _current_user: User = Depends(get_current_user),
) -> ChannelConfigResponse:
    """Return server-level channel configuration."""
    return _build_channel_config_response()


@router.put("/user/channels/config", response_model=ChannelConfigResponse)
async def update_channel_config(
    body: ChannelConfigUpdate,
    current_user: User = Depends(get_current_user),
) -> ChannelConfigResponse:
    """Update server-level channel configuration."""
    raw_updates = {k: v for k, v in body.model_dump(exclude_unset=True).items() if v is not None}
    if not raw_updates:
        raise HTTPException(status_code=400, detail="No fields to update")

    # The UI re-submits ``MASK`` ("********") for unchanged secret fields;
    # strip those so we don't overwrite real secrets with the sentinel.
    updates = strip_unchanged_secrets(raw_updates)
    if not updates:
        raise HTTPException(status_code=400, detail="No fields to update")

    # Enforce single Telegram chat ID (no comma-separated lists).
    chat_id = updates.get("telegram_allowed_chat_id", "")
    if chat_id and chat_id != "*" and "," in chat_id:
        raise HTTPException(
            status_code=422,
            detail="Only a single Telegram user ID is allowed. Remove commas.",
        )

    try:
        update_settings(updates)
    except ValueError as exc:
        raise HTTPException(status_code=422, detail=str(exc)) from exc

    await get_settings_store().save(updates, actor_user_id=current_user.id)

    reset_channel_clients(updates)

    return _build_channel_config_response()


@router.get("/user/channels/routes", response_model=ChannelRouteListResponse)
async def get_channel_routes(
    current_user: User = Depends(get_current_user),
    db: AsyncSession = Depends(get_async_db),
) -> ChannelRouteListResponse:
    """Return the current user's channel routes with enabled status."""
    routes = (
        (
            await db.execute(
                select(ChannelRoute)
                .where(ChannelRoute.user_id == current_user.id)
                .order_by(ChannelRoute.created_at)
            )
        )
        .scalars()
        .all()
    )
    return ChannelRouteListResponse(
        routes=[
            ChannelRouteResponse(
                channel=r.channel,
                channel_identifier=r.channel_identifier,
                enabled=r.enabled,
                created_at=r.created_at.isoformat() if r.created_at else "",
                last_inbound_at=r.last_inbound_at.isoformat() if r.last_inbound_at else None,
            )
            for r in routes
        ]
    )


@router.patch("/user/channels/routes/{channel}", response_model=ChannelRouteResponse)
async def update_channel_route(
    channel: str,
    body: ChannelRouteUpdate,
    current_user: User = Depends(get_current_user),
    db: AsyncSession = Depends(get_async_db),
) -> ChannelRouteResponse:
    """Toggle enabled status for a channel route.

    Single-channel enforcement: when enabling a channel, all other
    non-webchat routes for this user are automatically disabled so
    exactly one messaging channel is active at a time.

    If the user has no route and no known identifier for this channel yet
    (fresh onboarding), the selection is persisted via ``preferred_channel``
    only. The route row is created later when the identifier arrives,
    either via an inbound message (OSS) or via an explicit link call
    (premium). This keeps placeholder rows out of the database and avoids
    leaking the user's internal UUID into identifier-shaped UI fields.
    """
    from backend.app.channels import get_channel

    if channel == "webchat":
        raise HTTPException(status_code=400, detail="Webchat is always enabled")
    try:
        get_channel(channel)
    except KeyError as exc:
        raise HTTPException(status_code=404, detail=f"Unknown channel: {channel}") from exc

    # Re-query user in the current session to avoid detached instance issues
    user = await get_or_404_async(db, User, detail="User not found", id=current_user.id)

    if body.enabled:
        # Single-channel enforcement: disable all other non-webchat routes
        await db.execute(
            update(ChannelRoute)
            .where(
                ChannelRoute.user_id == user.id,
                ChannelRoute.channel != channel,
                ChannelRoute.channel != "webchat",
            )
            .values(enabled=False)
        )

    route = (
        await db.execute(select(ChannelRoute).filter_by(user_id=user.id, channel=channel))
    ).scalar_one_or_none()
    if route is None and user.channel_identifier:
        route = ChannelRoute(
            user_id=user.id,
            channel=channel,
            channel_identifier=user.channel_identifier,
            enabled=body.enabled,
        )
        db.add(route)
    elif route is not None:
        route.enabled = body.enabled

    if body.enabled:
        user.preferred_channel = channel
    else:
        await realign_preferred_channel_async(db, user)

    await db.commit()
    if route is not None:
        await db.refresh(route)
        return ChannelRouteResponse(
            channel=route.channel,
            channel_identifier=route.channel_identifier,
            enabled=route.enabled,
            created_at=route.created_at.isoformat() if route.created_at else "",
            last_inbound_at=route.last_inbound_at.isoformat() if route.last_inbound_at else None,
        )
    return ChannelRouteResponse(
        channel=channel,
        channel_identifier="",
        enabled=body.enabled,
        created_at="",
        last_inbound_at=None,
    )


@router.get("/channels/telegram/bot-info", response_model=TelegramBotInfoResponse)
async def get_telegram_bot_info(
    _current_user: User = Depends(get_current_user),
) -> TelegramBotInfoResponse:
    """Return the Telegram bot username, auto-discovered via getMe."""
    if not settings.telegram_bot_token:
        raise HTTPException(status_code=404, detail="No Telegram bot token configured")

    try:
        from backend.app.channels import get_channel
        from backend.app.channels.telegram import TelegramChannel

        channel = get_channel("telegram")
        if not isinstance(channel, TelegramChannel):
            raise HTTPException(status_code=404, detail="Telegram channel not available")

        me = await channel.bot.get_me()
        username = me.username or ""
        if not username:
            raise HTTPException(status_code=404, detail="Bot username not available")

        return TelegramBotInfoResponse(
            bot_username=username,
            bot_link=f"https://t.me/{username}",
        )
    except HTTPException:
        raise
    except Exception as exc:
        raise HTTPException(status_code=502, detail=f"Failed to fetch bot info: {exc}") from exc


# ---------------------------------------------------------------------------
# Model config
# ---------------------------------------------------------------------------


def _build_model_config_response() -> ModelConfigResponse:
    return ModelConfigResponse(
        llm_provider=settings.llm_provider,
        llm_model=settings.llm_model,
        llm_api_base=settings.llm_api_base,
        vision_model=settings.vision_model,
        vision_provider=settings.vision_provider,
        heartbeat_model=settings.heartbeat_model,
        heartbeat_provider=settings.heartbeat_provider,
        compaction_model=settings.compaction_model,
        compaction_provider=settings.compaction_provider,
        reasoning_effort=settings.reasoning_effort,
    )


@router.get("/user/model/config", response_model=ModelConfigResponse)
async def get_model_config(
    _current_user: User = Depends(get_current_user),
) -> ModelConfigResponse:
    """Return server-level LLM model configuration."""
    return _build_model_config_response()


@router.put("/user/model/config", response_model=ModelConfigResponse)
async def update_model_config(
    body: ModelConfigUpdate,
    current_user: User = Depends(get_current_user),
) -> ModelConfigResponse:
    """Update server-level LLM model configuration.

    NOTE: In single-tenant (OSS) mode, all authenticated users are
    effectively admins and can modify these settings. The premium layer
    adds role-based guards via AdminConfigGuardMiddleware to restrict
    this to admin users only.
    """
    updates = {k: v for k, v in body.model_dump(exclude_unset=True).items() if v is not None}
    if not updates:
        raise HTTPException(status_code=400, detail="No fields to update")

    try:
        update_settings(updates)
    except ValueError as exc:
        raise HTTPException(status_code=422, detail=str(exc)) from exc

    await get_settings_store().save(updates, actor_user_id=current_user.id)
    return _build_model_config_response()


# ---------------------------------------------------------------------------
# Provider / model enumeration
# ---------------------------------------------------------------------------


@router.get("/user/providers", response_model=list[ProviderInfo])
async def list_providers(
    _current_user: User = Depends(get_current_user),
) -> list[ProviderInfo]:
    """List available LLM providers from any-llm."""
    return get_configured_providers()


@router.get("/user/providers/{provider}/models")
async def list_provider_models(
    provider: str,
    api_base: str | None = Query(None),
    _current_user: User = Depends(get_current_user),
) -> list[str]:
    """List available models for a provider."""
    try:
        return await get_models(provider, api_base=api_base)
    except Exception as exc:
        raise HTTPException(status_code=502, detail=f"Failed to list models: {exc}") from exc


# ---------------------------------------------------------------------------
# Heartbeat logs
# ---------------------------------------------------------------------------


@router.get("/user/heartbeat-logs", response_model=HeartbeatLogListResponse)
async def get_heartbeat_logs(
    limit: int = Query(50, ge=1, le=200),
    current_user: User = Depends(get_current_user),
    db: AsyncSession = Depends(get_async_db),
) -> HeartbeatLogListResponse:
    """List heartbeat logs for the current user, most recent first."""
    total: int = (
        await db.scalar(
            select(sa_func.count(HeartbeatLog.id)).where(HeartbeatLog.user_id == current_user.id)
        )
    ) or 0

    logs = (
        (
            await db.execute(
                select(HeartbeatLog)
                .where(HeartbeatLog.user_id == current_user.id)
                .order_by(HeartbeatLog.created_at.desc())
                .limit(limit)
            )
        )
        .scalars()
        .all()
    )

    return HeartbeatLogListResponse(
        total=total,
        items=[
            HeartbeatLogItemResponse(
                id=log.id,
                user_id=log.user_id,
                action_type=log.action_type or "send",
                message_text=log.message_text or "",
                channel=log.channel or "",
                reasoning=log.reasoning or "",
                tasks=log.tasks or "",
                created_at=log.created_at.isoformat() if log.created_at else "",
            )
            for log in logs
        ],
    )


@router.delete("/user/heartbeat-logs", response_model=DeleteHeartbeatLogsResponse)
async def delete_heartbeat_logs(
    current_user: User = Depends(get_current_user),
    db: AsyncSession = Depends(get_async_db),
) -> DeleteHeartbeatLogsResponse:
    """Delete all heartbeat logs for the current user."""
    deleted: int = cast(
        "CursorResult[object]",
        await db.execute(
            delete(HeartbeatLog)
            .where(HeartbeatLog.user_id == current_user.id)
            .execution_options(synchronize_session="fetch")
        ),
    ).rowcount
    await db.commit()
    return DeleteHeartbeatLogsResponse(status="deleted", deleted=deleted)


# ---------------------------------------------------------------------------
# LLM usage summary
# ---------------------------------------------------------------------------


@router.get("/user/llm-usage", response_model=LLMUsageSummary)
async def get_llm_usage(
    days: int = Query(30, ge=1, le=365),
    current_user: User = Depends(get_current_user),
    db: AsyncSession = Depends(get_async_db),
) -> LLMUsageSummary:
    """Aggregate LLM usage for the current user over the last N days."""
    since = datetime.datetime.now(datetime.UTC) - datetime.timedelta(days=days)

    rows = (
        await db.execute(
            select(
                LLMUsageLog.purpose,
                sa_func.count(LLMUsageLog.id).label("call_count"),
                sa_func.coalesce(sa_func.sum(LLMUsageLog.input_tokens), 0).label(
                    "total_input_tokens"
                ),
                sa_func.coalesce(sa_func.sum(LLMUsageLog.output_tokens), 0).label(
                    "total_output_tokens"
                ),
                sa_func.coalesce(sa_func.sum(LLMUsageLog.total_tokens), 0).label("total_tokens"),
                sa_func.coalesce(sa_func.sum(LLMUsageLog.cost), 0).label("total_cost"),
            )
            .where(
                LLMUsageLog.user_id == current_user.id,
                LLMUsageLog.created_at >= since,
            )
            .group_by(LLMUsageLog.purpose)
        )
    ).all()

    by_purpose = [
        LLMUsageByPurpose(
            purpose=row.purpose or "",
            call_count=int(row.call_count),
            total_input_tokens=int(row.total_input_tokens),
            total_output_tokens=int(row.total_output_tokens),
            total_tokens=int(row.total_tokens),
            total_cost=float(row.total_cost),
        )
        for row in rows
    ]

    return LLMUsageSummary(
        total_calls=sum(p.call_count for p in by_purpose),
        total_tokens=sum(p.total_tokens for p in by_purpose),
        total_cost=sum(p.total_cost for p in by_purpose),
        by_purpose=by_purpose,
    )