# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Clawbolt",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/user/profile": {
      "get": {
        "summary": "Get Profile",
        "description": "Return the current user's profile.",
        "operationId": "get_profile_api_user_profile_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfileResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update Profile",
        "description": "Partial update of the current user's profile.",
        "operationId": "update_profile_api_user_profile_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserProfileUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfileResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/user/data-sharing-consent": {
      "get": {
        "summary": "Get Data Sharing Consent",
        "description": "Return the current user's data sharing consent state.",
        "operationId": "get_data_sharing_consent_api_user_data_sharing_consent_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataSharingConsentResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update Data Sharing Consent",
        "description": "Toggle the current user's data sharing consent.\n\nStamps ``data_sharing_consent_at`` with the current UTC time on every\ncall, regardless of whether the value changed. This makes the column\na \"last toggled at\" timestamp rather than a \"first opted in at\" one,\nwhich is the cheaper guarantee to keep correct: a one-shot accidental\ndouble-PUT can't drift the timestamp.",
        "operationId": "update_data_sharing_consent_api_user_data_sharing_consent_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DataSharingConsentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataSharingConsentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/user/channels/config": {
      "get": {
        "summary": "Get Channel Config",
        "description": "Return server-level channel configuration.",
        "operationId": "get_channel_config_api_user_channels_config_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChannelConfigResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update Channel Config",
        "description": "Update server-level channel configuration.",
        "operationId": "update_channel_config_api_user_channels_config_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChannelConfigUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChannelConfigResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/user/channels/routes/{channel}": {
      "patch": {
        "summary": "Update Channel Route",
        "description": "Toggle enabled status for a channel route.\n\nSingle-channel enforcement: when enabling a channel, all other\nnon-webchat routes for this user are automatically disabled so\nexactly one messaging channel is active at a time.\n\nIf the user has no route and no known identifier for this channel yet\n(fresh onboarding), the selection is persisted via ``preferred_channel``\nonly. The route row is created later when the identifier arrives,\neither via an inbound message (OSS) or via an explicit link call\n(premium). This keeps placeholder rows out of the database and avoids\nleaking the user's internal UUID into identifier-shaped UI fields.",
        "operationId": "update_channel_route_api_user_channels_routes__channel__patch",
        "parameters": [
          {
            "name": "channel",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Channel"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChannelRouteUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChannelRouteResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/user/model/config": {
      "get": {
        "summary": "Get Model Config",
        "description": "Return server-level LLM model configuration.",
        "operationId": "get_model_config_api_user_model_config_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelConfigResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update Model Config",
        "description": "Update server-level LLM model configuration.\n\nNOTE: In single-tenant (OSS) mode, all authenticated users are\neffectively admins and can modify these settings. The premium layer\nadds role-based guards via AdminConfigGuardMiddleware to restrict\nthis to admin users only.",
        "operationId": "update_model_config_api_user_model_config_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModelConfigUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelConfigResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/user/providers/{provider}/models": {
      "get": {
        "summary": "List Provider Models",
        "description": "List available models for a provider.",
        "operationId": "list_provider_models_api_user_providers__provider__models_get",
        "parameters": [
          {
            "name": "provider",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Provider"
            }
          },
          {
            "name": "api_base",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Api Base"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  },
                  "title": "Response List Provider Models Api User Providers  Provider  Models Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/user/heartbeat-logs": {
      "get": {
        "summary": "Get Heartbeat Logs",
        "description": "List heartbeat logs for the current user, most recent first.",
        "operationId": "get_heartbeat_logs_api_user_heartbeat_logs_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HeartbeatLogListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Heartbeat Logs",
        "description": "Delete all heartbeat logs for the current user.",
        "operationId": "delete_heartbeat_logs_api_user_heartbeat_logs_delete",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteHeartbeatLogsResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DataSharingConsentRequest": {
        "properties": {
          "consent": {
            "type": "boolean",
            "title": "Consent"
          }
        },
        "type": "object",
        "required": [
          "consent"
        ],
        "title": "DataSharingConsentRequest",
        "description": "Body for ``PUT /api/user/data-sharing-consent``.\n\nSingle boolean. The endpoint always stamps ``data_sharing_consent_at``\nwith ``now()`` regardless of whether ``consent`` is ``True`` or\n``False``, so consent toggle history can be reconstructed even when\nno separate audit table exists."
      },
      "DataSharingConsentResponse": {
        "properties": {
          "data_sharing_consent": {
            "type": "boolean",
            "title": "Data Sharing Consent"
          },
          "data_sharing_consent_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data Sharing Consent At"
          }
        },
        "type": "object",
        "required": [
          "data_sharing_consent",
          "data_sharing_consent_at"
        ],
        "title": "DataSharingConsentResponse",
        "description": "Returned by the consent setter and getter.\n\n``data_sharing_consent_at`` is the timestamp of the last toggle,\nnot \"first opted in.\" If a user opts in, then opts out, this column\nholds the opt-out time."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
    }
  }
}